+++ examples/use-cases/azure-runner-fleet/providers.tf
@@ -1,0 +1,9 @@
+provider "azurerm" {
+  subscription_id = var.azure_subscription_id
+
+  features {}
+}
+
+provider "daytona" {
+  api_url = var.daytona_api_url
+}